feat: improve error handling
In previous versions all problems were show-stopping errors. Now the
tool tracks both warnings and errors. Warnings indicate a problem that
does not require the tool to stop. For example, if there is a problem
install a single plugin, the tool can still try to install other
plugins. Errors, however, cause the tool to stop further work
immediately. For example, if the tool cannot create a directory to store
plugins, then it should not do anything else.

I also fixed a logical error in the `reconcile` function in plugin.go.
Previously, I was checking whether `statesByName` was nil. This was
a mistake for two reasons. First, and most importantly, the check would
never be true because (in `discoverPlugins`) I initialize the map to be
non-nil. Second, it's not clear to me that I should care if the map is
empty. That may only indicate that no plugins are installed yet. In that
case, the tool should install everything in the user's configuration
file.

Finally, I updated the TODO file by removing one item (about errors and
warnings) and clarifying the things I need to do to improve the code in
internal/git.