fix: silence git while pluggo works
When using a machine without a personal access token from GitHub, git
kept trying to get a username and password from me even as the spinner
spun. This is...bad.

The solution here is simply to silence git. The user will be told about
failures, and they will have to figure things out after pluggo finishes.